renamed component for create club form
